350.3. More on Using Portals(Section 25: Using React Portals)
We can use a feature called z index a z index is essentially a CSSA rule that's going to control what

element gets rendered on top of another element we assign z index properties integer values anywhere

from negative values all the way to positive values.

The higher a element or Seamew the higher z index value that an element has makes it more likely or

higher priority to be displayed on the top of the screen.

So right now if we give our model a z index of 10 it will be displayed on top of the sidebar which by

default has a z index of 0 or technically undefined.

In this case.

So to fix this up I'm going to go back over find my modal class right here and I'll give it a Z dash

index of 10.

Like so.

So now this has a higher z index.

They on our Sidebar.

So our motile should be displayed on top.

So I'm going to save that I'll flip back over again or refresh.

And now we see that the sidebar goes away.

It doesn't technically go away.

It's still in the background.

It's just that our gray background right here is visible on top of the sidebar.

OK.

So this looks like it's working out pretty well but I'm going to throw a little bit of a wrench into

this plan right here.

OK.

You got through a little problem in remember when ever we show content with react.

We are always showing all of our components inside of a single div like so and so I want you to imagine

that maybe at some point in time we put together a modal component which is exactly what we really want

to do for this daily screen or that delete page.

So remember this Modal right here is always going to be deeply nested inside of our application.

We might have a tremendous amount of H-G mail or a tremendous amount of GSX with a ton of styling as

parents to that modal component.

So this motile right here is going to be wrapped by maybe two or three divs inside of stream delete

maybe another div inside the app.

Another day with the idea of root right here and every one of those divs can have some amount of styling

being applied to it.

So I just want you to keep that in mind for one second.

When I threw a little wrench into our current example.

All right.

I'm going to go back over to my editor and I'm going to wrap my modal element right here with a new

div.

I'm going to give that div a class of positioned and then I'll add in a selector for it right here positioned

like so now I'm going to give positioned to very specific rules.

I'm going to say position is relative and I'm going to give it a z index of 0.

Like so Zangana saved this.

We're going to flip back or the browser and we're going to see what happens.

So when I refreshed the page we're back to the original case where I see I am a sidebar appearing on

the left hand side of the screen.

So why is that.

You'll notice that we still have motile right here with a z index of 10 which is far greater than our

sidebar which has a z index of essentially undefined.

So there is something about adding in positioned with a position of relative and z index of 0 that through

the styling off entirely.

So let's take a look at a diagram and understand why we're now seeing the sidebar on top.

So here we go right here.

So we now have our body element at the very top.

We still have the sidebar with the z index essentially of undefined on just going put 0 on there for

the sake of discussion.

We now have something called the positioned element right here that has that position of relative Z

next of 0.

And as a child 2 it is our motile with a z index of 10.

So here's the issue.

Whenever we place a div with a position of relative and a z index of some defined value even if it's

zero.

So we have some assigned value for z index.

It creates something called a stacking context.

There are many different ways to create a stacking context with success.

We can create a new stacking context by using a position of relative and a z index of 0.

But that's just one of many ways of doing this.

When we create a stacking context it essentially creates a new way of comparing different sibling elements

that have assigned z index values that might want to be rendered on top of each other.

Essentially when we create this stacking context right here we no longer compare sidebars z index with

a modals z index.

Instead we compare these sidebars the index with the root element that has the stacking context that

contains the motile.

So now it is the sidebar with a z index of undefined against the positioned element with a z index of

zero.

And in this case sidebar wins effectively there is the indexes are the same but the sidebar element

occurs later.

See how we have positioned right here and then sidebar is written out.

There is an element listed later on inside of the body element if these indexes are the same then whichever

element occurs later or second will be displayed on top of the other.

So that's why the sidebar is now being displayed on top because we are creating this new stacking context.

And so we are now comparing the sidebar against positioned all right and so we end up with this bad

example right here.

So how would we fix this.