350.4. More on Using Portals(Section 25: Using React Portals)
Well before we even think about fixing it you might say to me.

All right.

Well Stephen this seems like a really silly example.

But we have to do is take out this div with class positioned right.

That seems like a real simple example or maybe we should change the z index a position right here.

Now we could change it to like 99.

And now that's going to definitely beat out the sidebar.

So if I save this now and refresh we're back to the original good case.

Well here's the issue.

Remember I've now made a pretty big deal saying that when we have a re-act application we've got all

of these different divs all these different elements all these different components.

And our model might be deeply nested inside there.

And so this is a kind of contrived example of how a application might look that you are going to work

on some day you might work on an application that already has hundreds of components inside of it and

you might have a super deeply nested component structure.

And so it's not always going to be quite so easy for you to just say oh yeah I just need to like change

some parents DSS because this might be CSSA that has been super finely tuned and if you change it you

might break the entire application.

So it's not always a good solution to just say oh yeah just delete that div right there and you'll be

good to go or to change the z index on it and you'll be good to go.

It's incredibly likely that if you're adding a motile into some existing application you might end up

having to break the entire layout and rewrite all the CSSA for the project.

At least the layout us to get a modal working.

So just saying oh yeah I'm going to change the styling or the z index on this position.

Elmet right here or delete it entirely is not a good solution.

It's not a solution at all.