350.5. More on Using Portals(Section 25: Using React Portals)
So again how are we going to do this.

Well essentially we're going to use a little bit of a cheat.

Here's what we're going to do rather than showing our model as a child of positions.

We're going to instead show the modal as a child of the body.

Now you might look at this and still think Stephen way that goes against what you were just saying two

seconds ago.

Well not quite.

Remember when we look at this diagram right here of how we have to put re-act applications together

we had said that if the stream delete wants to show a modal the modal component will be rendered down

here.

So I'm essentially saying rather than follow the normal rules of re-act where we nest all of our components

underneath each other we should try to kind of like break or bend this rule a little bit and that's

exactly what we're going to do by using this portable thing with portal's.

We don't have to stick with this type of component hierarchy where everything is a child of this div

with the ID of root.

Instead when we use a portal we can essentially say OK stream delete We want you to render a modal component

but we don't want to have you render a modal as a direct child.

Instead try to render the model component but make it a child of some other element in our HMO hierarchy

such as the body element way back up here.

So that's the core purpose of a portal.

A portal allows us to render some element not as a direct child.

We can instead render that element or that component as a child of some other element inside our HDL

structure.

Most commonly the body why would we want to do this at all.

Well we would do this to get around all this rules of stacking contexts or trying to show a modal or

CSSA issues or stuff like that.

There are many different scenarios.

Modals are just one of them but it's the easiest one to show where we might want to render some element

or some component as a child of some other element inside of our application as opposed to just a direct

child of some given component.

So now if we render the modal as a direct child of the body we can set any z index on it that we want.

We will not run into that same stacking context issue that we saw a second ago and will be totally fine

and the modal will show up correctly and not conflict with any other element inside of application.

And so if we go back here to this HTL example the Senshi we would imagine still leaving position there

and leaving whatever other content position had you know maybe position had a bunch of text inside it

or something like that.

And we're going to instead just take our div with class modal and make it a direct child of the body

like so.

So now if I save this we can go back over and we're back to good operation as we saw before.

OK.

That's that's pretty much it.

That's why we make use of portal's.

Now the one part I left out here is telling you how we use a portal.

So let's take a quick pause right here.

When we come back in in the next section I'll show you how we create a portal and render something into

it.

So quick break in I'll see you in just a minute.