352.1. Hiding a Modal(Section 25: Using React Portals)
In last video we got our model to appear on the screen.

But of course it really does not look that good.

So in this video we're going to focus on adding some better styling to this thing and also getting the

ability to customize our model when it appears I want to first show you the semantic documentation around

modals just to give you a better idea of what we're going to do.

So in a new browser tab I'm going to go to a semantic dash UI dot com once every year all select menu

and then I'll go down to the very bottom find the modules section and will find the modal right there.

Now the default modal include include with semantic UI expects that we are using Jay query and we are

of course not using Jake where we are using entirely react.

So essentially we are kind of using the styling of modals from semantic UI but we're not making use

of any of the built in functionality or the javascript side of modals as they are included in semantic

UI.

So on this page I can scroll down a little bit and I just want to show you this example right here for

actions.

So inside of a modal we can place a div with the class of Hetter a div with the class content and a

div with class actions.

Those three divs right there generally line up with this mockup right here.

So at the very top is the div with Hetter whatever show is shown in the middle is the div with content.

And then finally at the bottom is different class actions.

That's going to be whatever show is on the bottom down here.

Let's just try to write these divs in and get some better styling going on for right now.

So back inside my editor I'm going to delete the jibberish I had inside there and then I will place

a div with the class name of Hetter and I'll give it some text like let's say delete stream.

That'll work.

I'll put a div in here with the class name of content and I'll give it some text like Are you sure you

want to delete the stream.

And then finally a div with the class name of actions and that will be that to do the delete and the

cancel button at the bottom.

So class name of actions.

Let's get a fix there.

And inside that div I'll do a button with the class name of UI primary arsole let's drop off the primary

just replace it with button like so that will give us a nice gray button for cancell and then for the

delete button we can do a class name you my primary button and we'll give that one the text delete like

so.

Now let's save this and see how this looks now.

So flip back over and that looks way better than it did before so I got the nice header on there.

I've got the nice content and the delete text right there as well.

So this definitely looks like a pretty good improvement but there's just one little issue here.

There always is an issue right when is there never an issue.

I want you to notice how right now if I click off of the modal like on the background over here the

model just stays on the screen.

Now typically whenever we have a modal window if the user clicks in the background we're going to want

to dismiss the modal window entirely and just hide it or go do something else.

Essentially we do not want to lock a user into having to click on these two buttons right here.