352.2. Hiding a Modal(Section 25: Using React Portals)
I think you can agree with me that you probably feel like if you hit Baqir or on the background you

want to see the modal go away.

So let's figure out how he could do that.

All right so I'm going to flip back over.

Remember that right now we're only showing this motile on the screen.

If the user is out the route streams slash delete.

So a very easy way that we could get this motile to go away is to just navigate to a different route

inside of our application.

So essentially if a user clicks on the background div we want to navigate the user to some other location

or path inside of our app.

We can get a user to navigate programmatically by making use of that history object that we created

a little bit ago.

So inside my motile at the very top I'm going to import history from up one directory history.

So this is remember that history that Jesus object or the history such as file where we created our

history for re-act router then on my div with Diemer modals visible active I can add on a on click event

handler So anytime that a user clicks on this thing we can call history dot push and we can go back

to our route route like so.

So notice how I wrapped this inside of a ero function.

So anytime we click on that thing run that arrow function which is going to call history don't push

it let's try saving this and give it a quick shot so I'll flip back over.

And now when I click on that I see the window go away.

Let's try that again one more time because it happened pretty darn quickly.

I'm going to again go to localhost three thousand streams slash delete.

And now if I click on say the background Yup I go back to my main stream window.

But let's do this one more time and I want to show you some interesting behavior this time around.

I'm going to click like just to the left of the delete button.

Let's say that our user comes in here and they try to click that delete button but they accidentally

clicked just off to the left side.

So if I click just to the left hand side that still dismisses the entire model or it navigates us away.

Well why is that.

Well essentially this is very basic event handling this is event propagation.

So remember this is default javascript behavior default HTL behavior.

If we ever trigger an event on some child elements and that child element Element does not handle that

event the event is going to essentially bubble up to some parent element on leave he gets caught with

an event handler.

So in this case if I click on any of the elements inside of any of these things that click event is

going to bubble up to this device.

It's going to run the on click event handler right here which is going to run history push.

So that's definitely something that we probably don't want to have happen.