352.3. Hiding a Modal(Section 25: Using React Portals)
So to make sure that doesn't occur we can at a click event handler to the window div for the content

div right here the day with the class name of UI standard modal.

So at in a click event handler to that as well.

And then inside if you're going to get the event argument that this function will be called with remember

whenever a event occurs we always get an event argument or event object as the argument.

And on that event argument we can call e or event eed for short stop propagation like so.

So stop propagation is going to make sure that that event does not continue to bubble up and go to this

div right here thus causing our window to get dismissed by navigating to some other page.

So I can save this and now we flip back over and again go manually to stream slash delete.

Now I can click all over this inner content window and nothing gets dismissed until I click on the background

behind it.

All right so that's a little bit better now.

Our model is coming along pretty well here but I want to point out a little bit of an issue at present

it has some hard coded content inside of here like hard coded delete stream and hard coded text of.

Are you sure you want to delete that stream.

Usually anytime we create a modal we definitely want to make sure that it is re-usable in nature.

So there might be many other locations inside of our application where we want to show a modal and in

every one of those cases we probably are going to want to have a different header different content

different buttons and we want different things to occur whenever a user clicks on those buttons.

In addition if a user dismisses the modal by clicking on that background in some cases we might not

want to navigate back to the root route.

So I think that we need to figure out a way to make this component far more usable than it currently

is.

So let's take a quick pause right here.

When we come back in the next section we'll do some refactoring to the modal.