354.1. React Fragments(Section 25: Using React Portals)
In last video we ran into a little bit of an issue when we tried to make our modal component a little

more reusable.

So I had wanted to pass in some variable called actions as a prop called actions and that was supposed

to assign some content or show some content inside of the div with the class name actions inside the

motile.

And so ideally this would be a set of buttons that a user can click on or something like that.

But we very quickly saw that when we wrapped those two buttons with a div semantic UI does not style

the model correctly.

And so this is what we end up with and it's all because we have that extra div in between our div with

class name actions and the two buttons.

So it specifically that div right there that is drawing the styling off.

As I mentioned in the last video that div is specifically there because we are not allowed to assign

multiple GSX elements like so to a single variable.

This is not valid javascript syntax you see right here.

And if you try to run this as a matter of fact like I'm going to save this file and I can flip back

over and it will even give us a very clear message like you're saying a Jasen or sibling elements must

be wrapped in an enclosing tag.

So that's what we're trying to do here we are trying to place some sibling elements next to each other.

But if we want to do so we have to wrap them with a single outside tag so we can have these two conflicting

requirements here on the one hand we have to have that div on the other hand.

We don't want the div because it throws off our styling.

So how are we going to solve this.

Well we're going to use a another little feature included inside re-act something called a re-act fragment

or react fragment is essentially a GSX looking element that is going to allow us to return multiple

elements or assigned multiple elements to a single variable.

But when it gets rendered onto the screen it doesn't actually produce any age to you Mel.

So you can think of Riak fragment as being like an invisible element that doesn't have any impact on

the Dom whatsoever.

So to make use of a fragment we can flip back over to stream delete and now in the dead of the div that

we had wrapping the two buttons a place a j a sex tackier and I'll say re-act fragment as the opening

tag.

And then re-act fragment as the closing tag like so.

And that's pretty much it.

So now we no longer have to GSX tags trying to be assigned to actions.

Instead we have a single tag being assigned to actions and a single tag contains two buttons internally.

So now we're not going to see that same syntax here that we had just a second ago.

In addition because Riak fragment does not actually produce any result or render element inside the

Dom when it gets rendered to the screen it's not going to throw off our styling as well.

So let's save this right here.

If I now flip back over I still see my motile on a screen and those two buttons are styled correctly

because I do not have that extra div being placed inside there.

As a matter of fact if you want to do something a bit interesting you can click on the element inspector

up here and then inspect that actions div..

So here's the div with class actions.

And if I expand it I see the two button elements so that re-act fragment did not render anything into

the DOM whatsoever.

OK.

That's pretty much it for re-act fragment.

Again you want to use this anytime you want to return multiple elements but not have some presence inside

the actual dom.