354.3. React Fragments(Section 25: Using React Portals)
The only reason that you would want to use re-act fragment is that there is some tooling out there like

some Linter some code quality checkers that do not think that empty tags like this are valid syntax.

So some code quality checkers that you might be running even inside of your own code editor will think

that this is invalid GSX and they will throw an error and say hey this is not valid.

You're doing something wrong here.

So the two approaches are identical they do the exact same thing.

The only reason that you would write out re-act fragment again is because some tooling thinks that this

is an error for the rest of this course.

If we ever use a fragment again I'm going to write out the long form that you see right here like re-act

fragment.

Just so you understand we are creating a fragment right there.

And in case you have any tooling that's throwing an error you're not going to see that error by writing

out the long form.

But if you want to absolutely feel free to do the shortened syntax like so OK so that's it for Riak

fragments.

Let's take a quick pause right here.

When we come back the next section we're going to continue to generalize our modal lesing do is make

sure that the on click callback right here can be customized by the parent component as well.